# nexus/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Nexus Control Plane API",
    "version": "2.2.0",
    "description": "Puente ChatGPT ↔ Nexus ↔ Cursor. POST /v1/prompt returns HTTP 202 immediately; poll GET /v1/tasks/{taskId} for results."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8787",
      "description": "Local control plane"
    }
  ],
  "paths": {
...
    "/v1/prompt": {
      "post": {
        "summary": "Enqueue a prompt for asynchronous execution",
        "description": "Returns HTTP 202 immediately with taskId and status QUEUED. The HTTP request never waits for Cursor or any executor.",
        "operationId": "enqueuePrompt",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["prompt"],
                "properties": {
                  "prompt": {
                    "type": "string"
                  },
                  "workspace": {
                    "type": "string",
                    "description": "Absolute workspace path for the task"
                  },
                  "metadata": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "client": {
                    "type": "object",
                    "properties": {
                      "request_id": {
                        "type": "string",
                        "description": "Idempotency key for duplicate prompt submissions"
                      }
                    }
                  },
                  "requestId": {
                    "type": "string",
                    "description": "Alias for client.request_id"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Task accepted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "taskId": { "type": "string", "format": "uuid" },
                    "status": { "type": "string", "enum": ["QUEUED"] },
                    "message": { "type": "string" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          }
        }
      }
    },
    "/v1/tasks": {
      "get": {
        "summary": "List tasks",
        "operationId": "listTasks",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["QUEUED", "WAITING_EXECUTOR", "RUNNING", "SUCCEEDED", "FAILED", "CANCELLED"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Task list"
          }
        }
      }
    },
...
    "/v1/recovery": {
      "post": {
        "summary": "Run a whitelisted recovery/repair action via executor.bootstrap",
        "description": "Protected by X-Nexus-Recovery-Token (or Authorization: Bearer). Does not depend on Cursor.",
        "operationId": "runRecovery",
        "security": [
          { "RecoveryToken": [] }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["action"],
                "properties": {
                  "action": {
                    "type": "string",
                    "enum": [
                      "repair-executor",
                      "register-executor",
                      "restart-worker",
                      "requeue-waiting",
                      "rollback-last-repair"
                    ]
                  },
                  "executorId": { "type": "string" },
                  "id": { "type": "string" },
                  "type": { "type": "string" },
                  "priority": { "type": "number" },
                  "enabled": { "type": "boolean" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Recovery action result" },
          "400": { "description": "Invalid action" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "NEXUS_API_KEY via Authorization: Bearer <key>"
      },
      "RecoveryToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Nexus-Recovery-Token"
      }
    }
  }
}